perf(tts): MPS 에서 bfloat16 사용
81d30d14 에서 float32 로 돌려놨는데, 마를렌(1.92B)을 올리다 bfloat16 이
통과하는 것을 보고 데비로도 재봤다. M4 맥미니 실측:

    데비 0.91B  float32   RTF 2.44
    데비 0.91B  bfloat16  RTF 1.97
    마를렌 1.92B bfloat16  RTF 2.36

float16 이 죽은 것은 정밀도가 아니라 표현 범위 때문이었다. bfloat16 은 지수
범위가 float32 와 같아 같은 자리를 통과하면서, 대역폭이 절반이라 더 빠르다 —
이 작업은 연산보다 가중치 운반이 병목이라 2배 큰 마를렌이 float32 데비보다
빠른 역전이 나온다.

메모리도 절반이라 마를렌이 float32 로는 7.7GB 로 16GB 기계 여유(약 6.5GB)를
넘던 것이 3.8GB 로 들어간다.

다만 이래도 실시간의 2배다. 디스코드 실시간 낭독을 이 기계로 옮기는 것은
아직 무리이고, 미리 만들어 쌓는 용도라면 쓸 만하다.

+++ run/services/tts/qwen3_tts_api.py
@@ -81,9 +81,11 @@ def init_device():
     if torch.cuda.is_available():
         device, dtype = "cuda:0", torch.bfloat16
     elif torch.backends.mps.is_available():
-        # float16 은 샘플링 단계에서 확률 텐서가 inf/nan 으로 넘친다(M4 실측:
-        # "probability tensor contains inf, nan or element < 0"). float32 라야 통과한다.
-        device, dtype = "mps", torch.float32
+        # float16 은 샘플링에서 확률 텐서가 inf/nan 으로 넘쳐 죽는다(M4 실측).
+        # bfloat16 은 지수 범위가 float32 와 같아 그 문제가 없고, 대역폭이 절반이라
+        # float32 보다 빠르다(데비 RTF 2.44 -> 1.97). 메모리도 절반이라 마를렌(1.92B)이
+        # 16GB 에 들어가는 것도 이 갈래 덕이다.
+        device, dtype = "mps", torch.bfloat16
     else:
         device, dtype = "cpu", torch.float32
